Refactor efend handlings
